# swagger_server/swagger/swagger.yaml
---
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Computation Lab"
  contact:
    email: "yshthdn@gmail.com"
host: "0.0.0.0:8080"
basePath: "/v1"
tags:
- name: "Computation modellig"
  description: "APIs to extrapolate ML data"
schemes:
- "https"
- "http"
paths:
  /compute_forecast:
    post:
      tags:
      - "Computation modelling"
      summary: ""
      description: ""
      operationId: "compute_forecast_post"
      parameters:
      - in: "body"
        name: "body"
        description: "Computation config needs to be added"
        required: true
        schema:
          $ref: "#/definitions/Config"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ComputeForecastResponse"
      x-swagger-router-controller: "swagger_server.controllers.computation_modelling_controller"
    put:
      tags:
      - "Computation modelling"
      summary: ""
      description: ""
      operationId: "compute_forecast_put"
      parameters:
      - in: "body"
        name: "body"
        description: "Pet object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/Config"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ComputeForecastResponse"
      x-swagger-router-controller: "swagger_server.controllers.computation_modelling_controller"
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Config:
    type: "object"
    properties:
      job_id:
        type: "integer"
        format: "int64"
      mem_request:
        type: "string"
      cpu_request:
        type: "string"
      time_request:
        type: "string"
        format: "date"
      project_id:
        type: "string"
      project_desc:
        type: "string"
      project_phase:
        type: "string"
      design_block_id:
        type: "string"
      design_block_desc:
        type: "string"
      tool_used:
        type: "string"
      tool_cmd:
        type: "string"
  ComputeForecastResponse:
    type: "object"
    properties:
      job_id:
        type: "integer"
        format: "int64"
      mem_request:
        type: "string"
      cpu_request:
        type: "string"
      time_request:
        type: "string"
        format: "date"
      project_id:
        type: "string"
      project_desc:
        type: "string"
      project_phase:
        type: "string"
      design_block_id:
        type: "string"
      design_block_desc:
        type: "string"
      tool_used:
        type: "string"
      tool_cmd:
        type: "string"
      predictions:
        type: "array"
        items:
          $ref: "#/definitions/Prediction"
    example:
      tool_used: "tool_used"
      time_request: "2000-01-23"
      project_id: "project_id"
      job_id: 0
      design_block_desc: "design_block_desc"
      design_block_id: "design_block_id"
      tool_cmd: "tool_cmd"
      project_phase: "project_phase"
      mem_request: "mem_request"
      project_desc: "project_desc"
      predictions:
      - cpu_predicted: "cpu_predicted"
        time_predicted: "time_predicted"
        mem_predicted: "mem_predicted"
        confidence_level: "LOW"
      - cpu_predicted: "cpu_predicted"
        time_predicted: "time_predicted"
        mem_predicted: "mem_predicted"
        confidence_level: "LOW"
      cpu_request: "cpu_request"
  Prediction:
    type: "object"
    properties:
      mem_predicted:
        type: "string"
      cpu_predicted:
        type: "string"
      time_predicted:
        type: "string"
      confidence_level:
        type: "string"
        enum:
        - "LOW"
        - "MEDIUM"
        - "HIGH"
    example:
      cpu_predicted: "cpu_predicted"
      time_predicted: "time_predicted"
      mem_predicted: "mem_predicted"
      confidence_level: "LOW"
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"